backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
requests
openai>=1.0.0
google-genai>=1.0.0
pillow>=11.1.0
email-validator
